+++ cli/devsecops_cli/__init__.py
@@ -2,7 +2,7 @@
 
 from importlib import import_module
 
-VERSION = "0.4.1"
+VERSION = "0.5.0"
 __version__ = VERSION
 
 __all__ = ["VERSION", "__version__", "main"]